+++ src/main/java/com/picksa/picksaserver/applicant/controller/ApplicantController.java
@@ -1,6 +1,6 @@
 package com.picksa.picksaserver.applicant.controller;
 
-import com.picksa.picksaserver.applicant.OrderCondition;
+import com.picksa.picksaserver.applicant.domain.OrderCondition;
 import com.picksa.picksaserver.applicant.dto.response.ApplicantAllResponse;
 import com.picksa.picksaserver.applicant.dto.response.ApplicantDetailResponse;
 import com.picksa.picksaserver.applicant.dto.response.ApplicantScheduleResponses;
