+++ src/main/java/com/picksa/picksaserver/applicant/domain/InterviewAvailableTime.java
@@ -1,4 +1,4 @@
-package com.picksa.picksaserver.applicant;
+package com.picksa.picksaserver.applicant.domain;
 
 import jakarta.persistence.Embeddable;
 import lombok.Getter;
