+++ src/main/java/com/picksa/picksaserver/applicant/domain/OrderCondition.java
@@ -1,4 +1,4 @@
-package com.picksa.picksaserver.applicant;
+package com.picksa.picksaserver.applicant.domain;
 
 public enum OrderCondition {
