+++ src/main/java/com/picksa/picksaserver/applicant/domain/Result.java
@@ -1,4 +1,4 @@
-package com.picksa.picksaserver.applicant;
+package com.picksa.picksaserver.applicant.domain;
 
 import lombok.Getter;
