+++ src/main/java/com/picksa/picksaserver/applicant/repository/AnswerRepository.java
@@ -1,7 +1,6 @@
 package com.picksa.picksaserver.applicant.repository;
 
-import com.picksa.picksaserver.applicant.AnswerEntity;
-import com.picksa.picksaserver.applicant.ApplicantEntity;
+import com.picksa.picksaserver.applicant.domain.AnswerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
