+++ src/main/java/com/picksa/picksaserver/applicant/repository/ApplicantQueryRepository.java
@@ -1,6 +1,6 @@
 package com.picksa.picksaserver.applicant.repository;
 
-import com.picksa.picksaserver.applicant.OrderCondition;
+import com.picksa.picksaserver.applicant.domain.OrderCondition;
 import com.picksa.picksaserver.applicant.dto.response.ApplicantResponse;
 import com.picksa.picksaserver.applicant.dto.response.ApplicantScheduleResponse;
 import com.picksa.picksaserver.global.domain.Part;
