+++ src/main/java/com/picksa/picksaserver/applicant/repository/InterviewScheduleRepository.java
@@ -1,6 +1,6 @@
 package com.picksa.picksaserver.applicant.repository;
 
-import com.picksa.picksaserver.applicant.InterviewScheduleEntity;
+import com.picksa.picksaserver.applicant.domain.InterviewScheduleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
