+++ src/main/java/com/picksa/picksaserver/evaluation/dto/request/DecideRequest.java
@@ -1,6 +1,6 @@
 package com.picksa.picksaserver.evaluation.dto.request;
 
-import com.picksa.picksaserver.applicant.Result;
+import com.picksa.picksaserver.applicant.domain.Result;
 
 public record DecideRequest (
     Result result
